add support for json schema
